Fix for tblImg()/getTokenImage() (and probably others) for webp images.
The raw asset change breaks several existing functions for webp (and
possibly other image types) when returning an asset to be displayed in
html/html5 views.

This change will always fetch the png version of an asset unless the raw=true
query flag is present. Unfortionately this indroduces a small compatability
issue but given the size of the compatability issue of not requiring the
explicit request for raw breaking other exisiting functions this is by far
the lesser of two evils.